Remove the size variable, replace with size().
This comes with a speed cut and I'm pretty sure doesn't change
the final file size at all. Oh well. On an unrelated note, it
turns out that I had forgetten to increase the size variable
whenever new cells are added, so it would always push_back the
cells if the index was over 10. So that's fixed now.