Load instructions before interpreting.
Instead of reading from the file directly, the
interpreter now loads all of the instructions
into a vector and loops through the vector to get
the commands. This is faster for two reasons: one,
the program does not have to keep reading from the
file, and two, it can remove all the non-commands
as it loads them in, removing the need for looping
over comments.