Convert to object-oriented programming.
This change causes the final project to be about
5 or 6 kilobytes larger, and tends to run programs
slower. However, I think that optimizing my code
for speed or storage made the code less clear and
uglier, with little benefit. After all, if I were
going for the fastest ever, I would write it in
C instead of C++. As such, I have decided to use
the features of C++ to their benefit to write
clean, readable, and more easily maintainable
code. This is not to say that the optimization
was worthless, however. After this whole ordeal,
the final executable, with classes, is about 10
kilobytes smaller and much more streamlined. The
way it is now, I think it strikes the best balance
between practicality, speed, and maintainability.