Add 'explicit' keywords.
I added the 'explicit' keyword to both the Instructions and CellSpace
classes. This way they can't be defined with an equals sign (e.g.
Instructions in = argv[1]) and instead have to be defined with
parentheses (Instructions in(argv[1])). I figured that made the most
sense when considering the way they are used.